+++ packages/stylesheet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-pdf/stylesheet
 
+## 6.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-pdf/types@2.9.1
+
 ## 6.1.0
 
 ### Minor Changes